+++ samples/cpp/aliked_lightglue.cpp
@@ -6,266 +6,13 @@
 // Demonstrates ALIKED + LightGlue matching and XFeat feature extraction.
 
 #include <opencv2/features.hpp>
-#include <opencv2/dnn.hpp>
 #include <opencv2/imgcodecs.hpp>
 #include <opencv2/imgproc.hpp>
 #include <opencv2/highgui.hpp>
-#include <algorithm>
-#include <cfloat>
-#include <cmath>
 #include <iostream>
 
 using namespace cv;
 using namespace std;
-
-namespace
-{
-static Mat toGray(const Mat& image)
-{
-    if (image.channels() == 1)
-        return image;
-
-    Mat gray;
-    if (image.channels() == 3)
-        cvtColor(image, gray, COLOR_BGR2GRAY);
-    else if (image.channels() == 4)
-        cvtColor(image, gray, COLOR_BGRA2GRAY);
-    else
-        CV_Error(Error::StsBadArg, "XFeat expects grayscale, BGR, or BGRA image");
-    return gray;
-}
-
-static Mat toNCHW(const Mat& blob, int channelsHint = -1)
-{
-    CV_Assert(blob.dims == 4);
-    if (channelsHint > 0 && blob.size[1] == channelsHint)
-        return blob;
-
-    const bool isNHWC = (channelsHint > 0) ? (blob.size[3] == channelsHint) : (blob.size[3] < blob.size[1]);
-    if (!isNHWC)
-        return blob;
-
-    const int H = blob.size[1], W = blob.size[2], C = blob.size[3];
-    int outSize[] = {1, C, H, W};
-    Mat out(4, outSize, CV_32F);
-    const float* src = blob.ptr<float>();
-    float* dst = out.ptr<float>();
-    for (int c = 0; c < C; ++c)
-    {
-        for (int y = 0; y < H; ++y)
-        {
-            for (int x = 0; x < W; ++x)
-            {
-                const int srcIdx = (y * W + x) * C + c;
-                const int dstIdx = c * H * W + y * W + x;
-                dst[dstIdx] = src[srcIdx];
-            }
-        }
-    }
-    return out;
-}
-
-static float sampleNearest(const Mat& map, float x, float y, int normW, int normH)
-{
-    CV_Assert(map.type() == CV_32F);
-    const int H = map.rows;
-    const int W = map.cols;
-    if (H <= 0 || W <= 0)
-        return 0.0f;
-    if (normW <= 1 || normH <= 1)
-        return map.at<float>(0, 0);
-
-    const float fx = x * static_cast<float>(W - 1) / static_cast<float>(normW - 1);
-    const float fy = y * static_cast<float>(H - 1) / static_cast<float>(normH - 1);
-    const int ix = std::max(0, std::min(W - 1, cvRound(fx)));
-    const int iy = std::max(0, std::min(H - 1, cvRound(fy)));
-    return map.at<float>(iy, ix);
-}
-
-static float sampleBilinear(const Mat& map, float x, float y, int normW, int normH)
-{
-    CV_Assert(map.type() == CV_32F);
-    const int H = map.rows;
-    const int W = map.cols;
-    if (H <= 0 || W <= 0)
-        return 0.0f;
-    if (W == 1 && H == 1)
-        return map.at<float>(0, 0);
-    if (normW <= 1 || normH <= 1)
-        return map.at<float>(0, 0);
-
-    float fx = x * static_cast<float>(W - 1) / static_cast<float>(normW - 1);
-    float fy = y * static_cast<float>(H - 1) / static_cast<float>(normH - 1);
-    fx = std::max(0.0f, std::min(fx, static_cast<float>(W - 1)));
-    fy = std::max(0.0f, std::min(fy, static_cast<float>(H - 1)));
-
-    const int x0 = cvFloor(fx);
-    const int y0 = cvFloor(fy);
-    const int x1 = std::min(x0 + 1, W - 1);
-    const int y1 = std::min(y0 + 1, H - 1);
-    const float dx = fx - x0;
-    const float dy = fy - y0;
-
-    const float v00 = map.at<float>(y0, x0);
-    const float v01 = map.at<float>(y0, x1);
-    const float v10 = map.at<float>(y1, x0);
-    const float v11 = map.at<float>(y1, x1);
-    return (1.f - dx) * (1.f - dy) * v00 +
-           dx * (1.f - dy) * v01 +
-           (1.f - dx) * dy * v10 +
-           dx * dy * v11;
-}
-
-static bool detectAndComputeXFeatPostprocess(const Mat& image, const Mat& mask, dnn::Net& net,
-                                             vector<KeyPoint>& keypoints, Mat& descriptors,
-                                             int maxKeypoints = 2000,
-                                             float detectionThreshold = 0.05f,
-                                             int inputSize = 640)
-{
-    keypoints.clear();
-    descriptors.release();
-    if (image.empty())
-        return false;
-
-    Mat gray = toGray(image);
-    const float scale = static_cast<float>(inputSize) /
-                        static_cast<float>(std::max(gray.cols, gray.rows));
-    const int resizedW = std::max(1, cvRound(gray.cols * scale));
-    const int resizedH = std::max(1, cvRound(gray.rows * scale));
-    Mat resized;
-    resize(gray, resized, Size(resizedW, resizedH), 0, 0, INTER_LINEAR_EXACT);
-
-    Mat padded = Mat::zeros(inputSize, inputSize, CV_8UC1);
-    resized.copyTo(padded(Rect(0, 0, resizedW, resizedH)));
-
-    Mat blob;
-    dnn::blobFromImage(padded, blob, 1.0 / 255.0, Size(inputSize, inputSize), Scalar(), false, false, CV_32F);
-    net.setInput(blob);
-    vector<Mat> outs;
-    net.forward(outs, {"output_feats", "output_keypoints", "output_heatmap"});
-    CV_Assert(outs.size() == 3);
-
-    Mat featBlob = toNCHW(outs[0], 64);
-    Mat kptBlob = toNCHW(outs[1], 65);
-    Mat relBlob = toNCHW(outs[2], 1);
-    CV_Assert(featBlob.dims == 4 && kptBlob.dims == 4 && relBlob.dims == 4);
-
-    const int featC = featBlob.size[1];
-    const int featH = featBlob.size[2];
-    const int featW = featBlob.size[3];
-    const int kptC = kptBlob.size[1];
-    const int kptH = kptBlob.size[2];
-    const int kptW = kptBlob.size[3];
-    CV_Assert(featC == 64);
-    CV_Assert(kptC >= 64);
-
-    Mat reliability(relBlob.size[2], relBlob.size[3], CV_32F, relBlob.ptr<float>());
-    Mat heatmap = Mat::zeros(kptH * 8, kptW * 8, CV_32F);
-    const float* kptPtr = kptBlob.ptr<float>();
-    const int kptHW = kptH * kptW;
-
-    for (int y = 0; y < kptH; ++y)
-    {
-        for (int x = 0; x < kptW; ++x)
-        {
-            const int offset = y * kptW + x;
-            float maxLogit = -FLT_MAX;
-            for (int c = 0; c < kptC; ++c)
-                maxLogit = std::max(maxLogit, kptPtr[c * kptHW + offset]);
-
-            float sumExp = 0.f;
-            float probs[64];
-            for (int c = 0; c < 64; ++c)
-            {
-                probs[c] = std::exp(kptPtr[c * kptHW + offset] - maxLogit);
-                sumExp += probs[c];
-            }
-            if (kptC > 64)
-                sumExp += std::exp(kptPtr[64 * kptHW + offset] - maxLogit);
-            if (sumExp <= 0.f)
-                continue;
-
-            for (int c = 0; c < 64; ++c)
-            {
-                const int dy = c / 8;
-                const int dx = c % 8;
-                heatmap.at<float>(y * 8 + dy, x * 8 + dx) = probs[c] / sumExp;
-            }
-        }
-    }
-
-    Mat localMax;
-    dilate(heatmap, localMax, getStructuringElement(MORPH_RECT, Size(5, 5)));
-
-    struct Candidate
-    {
-        Point2f ptPadded;
-        float score;
-    };
-    vector<Candidate> candidates;
-    candidates.reserve(4096);
-
-    for (int y = 0; y < heatmap.rows; ++y)
-    {
-        const float* hm = heatmap.ptr<float>(y);
-        const float* mx = localMax.ptr<float>(y);
-        for (int x = 0; x < heatmap.cols; ++x)
-        {
-            const float h = hm[x];
-            if (h <= detectionThreshold || h != mx[x])
-                continue;
-            const float xf = static_cast<float>(x);
-            const float yf = static_cast<float>(y);
-            const float score = sampleNearest(heatmap, xf, yf, inputSize, inputSize) *
-                                sampleBilinear(reliability, xf, yf, inputSize, inputSize);
-            if (score <= 0.f)
-                continue;
-            candidates.push_back({Point2f(xf, yf), score});
-        }
-    }
-
-    if (candidates.empty())
-        return true;
-
-    const int keep = maxKeypoints > 0 ? std::min(maxKeypoints, static_cast<int>(candidates.size()))
-                                      : static_cast<int>(candidates.size());
-    std::partial_sort(candidates.begin(), candidates.begin() + keep, candidates.end(),
-                      [](const Candidate& a, const Candidate& b) { return a.score > b.score; });
-    candidates.resize(keep);
-
-    descriptors.create(keep, 64, CV_32F);
-    keypoints.reserve(keep);
-    const float* featPtr = featBlob.ptr<float>();
-    const int featHW = featH * featW;
-
-    for (int i = 0; i < keep; ++i)
-    {
-        const float xp = candidates[i].ptPadded.x;
-        const float yp = candidates[i].ptPadded.y;
-        const float xOrig = xp / scale;
-        const float yOrig = yp / scale;
-        const int ix = cvFloor(xOrig);
-        const int iy = cvFloor(yOrig);
-        if (ix < 0 || iy < 0 || ix >= image.cols || iy >= image.rows)
-            continue;
-        if (!mask.empty() && mask.at<uchar>(iy, ix) == 0)
-            continue;
-
-        float* dst = descriptors.ptr<float>(static_cast<int>(keypoints.size()));
-        for (int c = 0; c < 64; ++c)
-        {
-            Mat channel(featH, featW, CV_32F, const_cast<float*>(featPtr + c * featHW));
-            dst[c] = sampleBilinear(channel, xp, yp, inputSize, inputSize);
-        }
-        Mat row(1, 64, CV_32F, dst);
-        normalize(row, row, 1.0, 0.0, NORM_L2);
-        keypoints.emplace_back(Point2f(xOrig, yOrig), 1.0f, -1.0f, candidates[i].score);
-    }
-
-    descriptors = descriptors.rowRange(0, static_cast<int>(keypoints.size())).clone();
-    return true;
-}
-} // namespace
 
 static int runXFeatExample(const String& imgPath, const String& xfeatModel, const String& outputPath)
 {
@@ -276,15 +23,10 @@ static int runXFeatExample(const String& imgPath, const String& xfeatModel, cons
         return -1;
     }
 
-    dnn::Net net = dnn::readNetFromONNX(xfeatModel);
-
+    Ptr<XFeat> xfeat = XFeat::create(xfeatModel, 2000, 0.05f, 640);
     vector<KeyPoint> keypoints;
     Mat descriptors;
-    Mat mask(img.size(), CV_8UC1, Scalar::all(255));
-    detectAndComputeXFeatPostprocess(img, mask, net, keypoints, descriptors,
-                                     /*maxKeypoints=*/2000,
-                                     /*detectionThreshold=*/0.05f,
-                                     /*inputSize=*/640);
+    xfeat->detectAndCompute(img, Mat(), keypoints, descriptors);
 
     Mat canvas;
     drawKeypoints(img, keypoints, canvas, Scalar(0, 255, 0),
